requirements.txt
asn1crypto==0.24.0
blinker==1.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
CurrencyConverter==0.13.9
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
lxml==4.3.3
nltk==3.4.1
numpy==1.16.4
oauthlib==3.0.1
pandas==0.24.2
psycopg2-binary==2.8.2
pycparser==2.19
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.2.0
scikit-learn==0.21.2
scipy==1.3.0
selenium==3.141.0
six==1.12.0
SQLAlchemy==1.3.4
SQLAlchemy-Utils==0.33.11
urllib3==1.25.3
URLObject==2.4.3
Werkzeug==0.15.2
